+++ tremor-script/src/registry.rs
@@ -308,9 +308,9 @@ impl FunctionError {
 /// Wrapper around a function
 pub struct TremorFnWrapper {
     /// Name of the module the function is in
-    module: String,
+    pub module: String,
     /// Name of the function
-    name: String,
+    pub name: String,
     /// Boxed dyn of the implementaiton
     fun: Box<dyn TremorFn>,
 }
@@ -949,21 +949,30 @@ impl Aggr {
     }
 }
 
-#[cfg(test)]
-pub use tests::fun;
+/// Test utility to grab a function from the registry
+pub fn fun<'event>(
+    m: &'event str,
+    f: &'event str,
+) -> impl Fn(&[&Value<'event>]) -> FResult<Value<'event>> {
+    let func = registry().find(m, f).cloned();
+
+    move |args: &[&Value]| -> FResult<Value> {
+        match &func {
+            Ok(function) => function.invoke(&EventContext::new(0, None), args),
+            Err(_) => Err(FunctionError::MissingFunction {
+                m: m.to_string(),
+                f: f.to_string(),
+            }),
+        }
+    }
+}
+
+// #[cfg(test)]
+// pub use tests::fun;
 #[cfg(test)]
 mod tests {
     use super::*;
     use simd_json::prelude::*;
-
-    // Test utility to grab a function from the registry
-    pub fn fun<'event>(m: &str, f: &str) -> impl Fn(&[&Value<'event>]) -> FResult<Value<'event>> {
-        let f = registry()
-            .find(m, f)
-            .expect("could not find function")
-            .clone();
-        move |args: &[&Value]| -> FResult<Value> { f.invoke(&EventContext::new(0, None), args) }
-    }
 
     #[test]
     pub fn fun_error_equality_checks() {